separate view models on files, correct views bindings
