+++ LinksStorage/Pages/GroupPage.xaml
@@ -39,7 +39,7 @@
                                         </Image.Source>
                                     </Image>
 
-                                    <Label Grid.Column="1" Text="{Binding Alias}" LineBreakMode="TailTruncation" VerticalOptions="Center" TextColor="White"/>
+                                    <Label Grid.Column="1" Text="{Binding Name}" LineBreakMode="TailTruncation" VerticalOptions="Center" TextColor="White"/>
 
                                     <Button Grid.Column="2" Text="edit" Style="{StaticResource IconBtnFreeSolid}"
                                             Command="{Binding Source={RelativeSource AncestorType={x:Type viewModels:GroupVM}}, Path=EditLinkCommand}"
