+++ LinksStorage/Pages/LinkEditPage.xaml
@@ -8,7 +8,7 @@
     <Grid RowDefinitions="*, Auto" Padding="20">
         <VerticalStackLayout>
             <Label Text="Alias"/>
-            <Entry Text="{Binding Alias}"/>
+            <Entry Text="{Binding Name}"/>
             <Label Text="Url"/>
             <Entry Text="{Binding Url}"/>
         </VerticalStackLayout>
