+++ LinksStorage/ViewModels/LinkCreateInfo.cs
@@ -1,0 +1,3 @@
+﻿namespace LinksStorage.ViewModels;
+
+public record LinkCreateInfo(int Group);
