+++ LinksStorage/ViewModels/NavigationRoutes.cs
@@ -1,0 +1,8 @@
+﻿namespace LinksStorage.ViewModels;
+
+public static class NavigationRoutes
+{
+    public const string Root = "root";
+    public const string Group = "group";
+    public const string LinkEditForm = "linkEditForm";
+}
