+++ LinksStorage/ViewModels/RootGroupVM.cs
@@ -142,157 +142,3 @@ public partial class RootGroupVM : ObservableObject, IDisposable
         _messenger.Unsubscribe<DataPersistenceOutbox, EditLink>(this, nameof(EditLink));
     }
 }
-
-[QueryProperty(nameof(GroupId), "group")]
-public sealed partial class GroupVM : RootGroupVM
-{
-    public GroupVM(IMessagingCenter messenger, IServiceScopeFactory scopeFactory) : base(messenger, scopeFactory)
-    {
-        messenger.Subscribe<DataPersistenceOutbox, CreatedLink>(this, nameof(CreatedLink), AddLink);
-
-    }
-
-    [RelayCommand]
-    private async Task AddLink()
-    {
-        await Shell.Current.GoToAsync(NavigationRoutes.LinkEditForm, new Dictionary<string, object>()
-        {
-            ["payload"] = new LinkCreateInfo(GroupId)
-        });
-    }
-
-    private void AddLink(DataPersistenceOutbox _, CreatedLink args)
-    {
-        if(args.GroupId != GroupId) return;
-        Links.Add(new(){Id = args.Id, Name = args.Name, Url = args.Url});
-    }
-
-    protected override void MarkLinkAsFavorite(DataPersistenceOutbox _, MarkedLinkAsFavorite args)
-    {
-        if(Links.FirstOrDefault(x=>x.Id == args.Id) is not {} entry) return;
-        entry.IsFavorite = true;
-    }
-
-    protected override void RemoveMarkLinkAsFavorite(DataPersistenceOutbox _, RemovedMarkLinkAsFavorite args)
-    {
-        if (Links.FirstOrDefault(x => x.Id == args.Id) is not { } entry) return;
-        entry.IsFavorite = false;
-    }
-
-    public override void Dispose()
-    {
-        _messenger.Unsubscribe<DataPersistenceOutbox, CreatedLink>(this, nameof(CreatedLink));
-        base.Dispose();
-    }
-}
-
-public partial class LinkEditVM : ObservableObject, IQueryAttributable
-{
-    private readonly IMessagingCenter _messenger;
-    private static readonly object SenderMark = new();
-    private int _group;
-    private bool _isFromRoot;
-    private int _id;
-
-    [ObservableProperty]
-    private bool _isNew;
-
-    [ObservableProperty]
-    private string _name;
-    [ObservableProperty]
-    private string _url;
-
-    public LinkEditVM(IMessagingCenter messenger)
-    {
-        _messenger = messenger;
-    }
-
-    public void ApplyQueryAttributes(IDictionary<string, object> query)
-    {
-        var payload = query["payload"];
-        if (payload is LinkCreateInfo createInfo)
-        {
-            _id = 0;
-            _isNew = true;
-            _group = createInfo.Group;
-            return;
-        }
-
-        if (payload is not LinkEditInfo editInfo) throw new InvalidOperationException("Link Edit VM need for edit or create links");
-
-        _id = editInfo.Id;
-        _group = editInfo.Group;
-        Name = editInfo.Name;
-        Url = editInfo.Url;
-        _isFromRoot = editInfo.IsFromRoot;
-    }
-
-    [RelayCommand]
-    private async Task Save()
-    {
-        if (_isNew)
-        {
-            _messenger.Send(SenderMark, nameof(CreateLink), new CreateLink(_name, _url, _group));
-        }
-        else
-        {
-            _messenger.Send(SenderMark, nameof(EditLink), new EditLink(_id, _name, _url, _group));
-        }
-
-        await Shell.Current.GoToAsync("..", new Dictionary<string, object>()
-        {
-            ["group"] = _isFromRoot ? null : _group
-        });
-    }
-}
-
-public static class NavigationRoutes
-{
-    public const string Root = "root";
-    public const string Group = "group";
-    public const string LinkEditForm = "linkEditForm";
-}
-
-public record LinkEditInfo(int Group, int Id, string Name, string Url)
-{
-    public bool IsFromRoot => Group == 1;
-}
-
-public partial class LinkInfo : ObservableObject
-{
-    [ObservableProperty] private int _id;
-    [ObservableProperty] private string _name;
-    [ObservableProperty] private string _url;
-    [ObservableProperty] private bool _isFavorite;
-
-    public LinkInfo() { }
-    public LinkInfo(LinkInfoData data)
-    {
-        _id = data.Id;
-        _name = data.Name;
-        _url = data.Url;
-    }
-    
-    public void Update(EditLink data) => Update(data.Name, data.Url);
-
-    private void Update(string name, string url)
-    {
-        Name = name;
-        Url = url;
-    }
-}
-
-public partial class GroupInfo : ObservableObject
-{
-    [ObservableProperty] private int _id;
-    [ObservableProperty] private string _name;
-
-    public GroupInfo() { }
-    public GroupInfo(GroupInfoData data)
-    {
-        _id = data.Id;
-        _name = data.Name;
-    }
-}
-
-public record LinkCreateInfo(int Group);
